api/requirements.txt
fastapi
uvicorn[standard]
clickhouse-connect~=0.9.2
python-dotenv
